Version 0.5.0 - Scalabilité améliorée grâce à de nouveaux components et features mineures
