+++ portfolio/src/components/Navbar.vue
@@ -1,5 +1,5 @@
 <script setup>
-import { ref, watch, defineEmits } from 'vue';
+import { ref, watch } from 'vue';
 
 const isExpanded = ref(false);
 const emit = defineEmits(['update-width']);
